make identnum lock-free
We don't require locking here, as the ident count only ever goes
up. Being racily overwritten is impossible (writes into identmap
are still lock-protected) and the worst that can happen is getting
one less ident if something has in the meantime added something.

That is generally fine, as the alternative is blocking on the
mutex, getting the number, and the identmap expanding after it
gets unblocked; effectively the same thing.